feat(task): list 和 check 空行回车删除
